Move gc_mock_provider fixture to conftest.py
Problem: gc_mock_provider was a pytest fixture defined inside gc_test.py
at line 98, violating the project convention that all fixtures must be in
conftest.py (per CLAUDE.md).

Fix: Moved the fixture to libs/mngr/imbue/mngr/api/conftest.py with the
required imports, and removed the definition from gc_test.py.
